fix: :bug: fix password visibility on reset page
