+++ client/src/components/Pages/register/PasswordField/PasswordField.module.scss
@@ -1,0 +1,21 @@
+@import "src/scss/colors";
+@import "src/scss/variables";
+@import "src/scss/mixins/text";
+@import "src/scss/mixins/rtl";
+@import "src/scss/mixins/unit";
+
+.container {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.progress {
+  width: 30%;
+  @include mr-ltr(u(4));
+}
+
+.score {
+  font-weight: bold;
+  @include text("very-small");
+}
