+++ client/src/components/Pages/register/PasswordField/index.ts
@@ -1,0 +1,3 @@
+import PasswordField from "./PasswordField";
+
+export default PasswordField;
